Move PreciseKeys to quasar.common.data
